Localize Open Sans font and merge its CSS into main
I downloaded the CSS that we were fetching from Google Fonts along with
the font file referenced in it and they are now served locally from the
_css directory.

The Open Sans CSS is still in a separate file, but it’s imported from
style.css, so there is one less link tag in the HTML head.

This has the advantage of letting this CSS work offline, and it also
makes no requests in the background to nosy services like Google’s. The
disadvantage is that with a large directory of documents using their own
copy of this CSS you will download the same ~600 kB of fonts over and
over.