Remove locally hosted Google font
Open Sans is a nice font, but if I am ever remotely accessing a
directory full of documents made with this template I would end up
downloading the same 600 kB of font files over and over. I switched
to a decent list of common sans serif fonts and remove the locally
hosted Open Sans files from Google Fonts.

I put a full copy of this template with notes in the README in
~/work/areas/web/experiments as a reference on how to locally host a
Google font.